% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsize.R
\name{estimation}
\alias{estimation}
\title{Estimation de la taille de population}
\usage{
estimation(X0, Y0)
}
\value{
la taille de population
}
\description{
Prend en parametres des coordonn??es g??ographique et retour la taille de population ?? ce lieudit
}
